pyproject.toml
[project]
name = "clip-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kanghyun Ryu", email = "khryu@kist.re.kr" }
]
requires-python = ">=3.12"
dependencies = [
  "albumentations>=2.0.8",
  "autorootcwd>=1.0.1",
  "clip",
  "datasets>=4.0.0",
  "dvclive>=3.48.5",
  "gradio>=5.44.1",
  "ipykernel>=6.30.1",
  "jupyter-ai[all]>=2.31.6",
  "jupyterlab>=4.4.7",
  "kaggle>=1.7.4.5",
  "lightning>=2.5.4",
  "matplotlib>=3.10.6",
  "numpy>=2.3.2",
  "opencv-python>=4.11.0.86",
  "pillow>=11.3.0",
  "streamlit>=1.49.1",
  "timm>=1.0.19",
  "torch==2.7.*",
  "torchvision==0.22.*",
  "tqdm>=4.67.1",
  "transformers>=4.56.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
clip = { git = "https://github.com/openai/CLIP.git", rev = "a1d071733d7111c9c014f024669f959182114e33" }

[[tool.uv.index]]
name = "pytorch-cu128"
url  = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
]

